requirements.txt
Flask==3.0.0
Flask-SQLAlchemy>=3.0.5
Flask-Login>=0.6.3
Flask-WTF>=1.2.0
WTForms>=3.1.0
python-dotenv>=1.0.0
MarkupSafe>=2.1.0
